Added input validation for the chatCoverTextInput (userID, limited to digits, between 1 and 10 characters) and TextInput(message, limited to punctuation and alphanumeric characters, up to 425 characters) fields. Added CSS styling for dynamically sized button groups (where the size of the group varies according to the varying amount of buttons it may contain).
Added JS support for inserting a group of buttons that were defined from the backend and sent inside an array of object where each object contains the text for the buttons and aditionally, some "hidden" text that is to be inserted in the chat window and sent as the actual message whenever that particular button is clicked. The "button" object may also contain an icon which is to be inserted along the "visible" button text.  Added support for an event used to send information to the backed, without inserting that informtation in the chat window. Currently, this event, "getTypeOfPerson" is exclusively used to pass information about the first character in the userID variable. This may or may not be used by the backend.

Also added JS support for inserting multiple messages that were sent simultaneously in one event, by the back end. For this purpose, a new parameter, "multipleMessagesSignal", was added to the "serverMessage" event, on both the event listeners and emitters. If this parameter is equal to "complex", it means that the "text" parameter is actually an array of text strings, and for each of these, the "appendMessage" function is called, to insert each message as an individual "message bubble" in the chat window. If the "multipleMessagesSignal" parameter is not equal to "complex", then the "serverMessage" event listener wil treat the "text" parameter, like a single string.

[JS] Added a new event, "startChat" used specifically for starting a new chat in the Xcally server. This was done as a bandaid fix over an issue in the backend where the Xcally server sends an unexpedted response in certain situations, when sending a message. The old clientMessage event listener on the backend used the Xcally server's response to map the socket to an interaction ID, which was sent in the Xcally's server response for each new message posted. Due to the bug, this response could sometimes miss this interaction ID parameter, so it was necessary to map the socket to an interaction ID only when starting a new chat, and not each time a message is submitted to the Xcally server.

[JS] Added support for "hidden Text" in buttons, now, if a the ".texto-botones" span inside a button div includes an element with the ".texto-oculto" selector, the "button click handler" will insert the text contained in that element, into the chat window, and send that to the backend as part of the text for a clientMessage event.

[JS] Now the text Input used for entering the userID, and the one used for entering a new message, will be held to the following rules: the userID text input may only contain digits, if not, the "start Chat" button will remain opaque and unusable, and pressing "Enter" in that input, after inserting some string which includes anything other than a digit, will not start a chat. Meanwhile, the message text input can only hold alphanumeric characters, and the following punctuation characters: . , ? ¿ ¡ ! ' "